requirements.txt
pyTelegramBotAPI==4.9.0
openai==0.26.1
googletrans==3.1.0a0
pandas==1.5.3
python-dotenv==0.19.0